User model finished

+++ src/main/java/com/boardgames/hub/user/UserService.java
@@ -1,12 +1,15 @@
 package com.boardgames.hub.user;
 
 import com.boardgames.hub.exception.DuplicateResourceException;
+import com.boardgames.hub.exception.ResourceNotFoundException;
 import com.boardgames.hub.user.dto.UserCreateRequest;
 import com.boardgames.hub.user.dto.UserResponse;
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 @Slf4j
 @Service
@@ -16,14 +19,14 @@ public class UserService {
 
     @Transactional
     public UserResponse createUser(UserCreateRequest request) {
-        log.info("Attempting to create a new user with email: {}", request.getEmail());
+        log.info("Creating user with email: {}", request.getEmail());
 
         if (userRepository.existsByEmail(request.getEmail())) {
-            log.warn("User not created - email already exists: {}", request.getEmail());
+            log.warn("Email already exists: {}", request.getEmail());
             throw new DuplicateResourceException("User with this email already exists");
         }
         if (userRepository.existsByNickname(request.getNickname())) {
-            log.warn("User not created - nickname already exists: {}", request.getNickname());
+            log.warn("Nickname already exists: {}", request.getNickname());
             throw new DuplicateResourceException("User with this nickname already exists");
         }
 
@@ -41,15 +44,38 @@ public class UserService {
         return mapToResponse(savedUser);
     }
 
+    @Transactional(readOnly = true)
+    public List<UserResponse> getUsers() {
+        log.info("Getting all users");
+
+        return userRepository.findAll()
+                .stream()
+                .map(this::mapToResponse)
+                .toList();
+    }
+
+    @Transactional(readOnly = true)
     public UserResponse getUser(Long id) {
+        log.info("Getting user with ID: {}", id);
+
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("User not found with an ID: " + id));
+                .orElseThrow(() -> {
+                    log.warn("User not found with ID: {}", id);
+                    return new ResourceNotFoundException("User not found with ID: " + id);
+                });
 
         return mapToResponse(user);
     }
 
     @Transactional
     public void deleteUser(Long id) {
+        log.info("Deleting user with ID: {}", id);
+
+        if (!userRepository.existsById(id)) {
+            log.warn("User not found with ID: {}", id);
+            throw new ResourceNotFoundException("User not found with ID: " + id);
+        }
+
         userRepository.deleteById(id);
     }
